requirements.txt
APScheduler==3.1.0
base58==0.2.2
boto3==1.3.1
botocore==1.4.13
docopt==0.6.2
docutils==0.12
dominate==2.2.0
enum34==1.1.4
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
flask-nav==0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
funcsigs==1.0.2
futures==3.0.5
google-api-python-client==1.5.0
hjson==1.5.4
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
lambda-packages==0.5.0
MarkupSafe==0.23
MySQL-python==1.2.5
oauth2client==2.0.2
oauthlib==1.1.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.12
tqdm==4.5.0
tweepy==3.5.0
tzlocal==1.2.2
Unidecode==0.4.19
uritemplate==0.6
visitor==0.1.2
Werkzeug==0.11.9
wsgi-request-logger==0.4.5
WTForms==2.1
zappa==0.17.6
